histoire_bilingues/requirements.txt
replicate
requests
Pillow
python-dotenv
requests
googletrans==4.0.0-rc1
groq
time
ebooklib

